Moved burnsim exporter to new system, made changing motor reset simulation results
