+++ app.py
@@ -32,6 +32,7 @@ class App(QApplication):
         self.importExportManager = uilib.importExportManager.ImportExportManager(self)
         self.preferencesManager.preferencesChanged.connect(self.importExportManager.setPreferences)
         self.simulationManager.newSimulationResult.connect(self.importExportManager.acceptSimRes)
+        self.fileManager.newMotor.connect(self.importExportManager.acceptNewMotor)
 
         if self.headless:
             if len(args) < 3:
