+++ uilib/converters/__init__.py
@@ -1,3 +1,4 @@
 from .burnsimImporter import BurnSimImporter
 
 from .engExporter import EngExporter
+from .burnsimExporter import BurnSimExporter
